feat: adapt strategy CR changes to new logic for consistency
